[GEOS-11312] Changed Tally memory calculation datatype from int to long to avoid possible overflow.

+++ src/wms/src/main/java/org/geoserver/wms/legendgraphic/BufferedImageLegendGraphicBuilder.java
@@ -289,7 +289,7 @@ public class BufferedImageLegendGraphicBuilder extends LegendGraphicBuilder {
 
         // final checks
         if (finalLegend == null) throw new IllegalArgumentException("no legend passed");
-        int maxMemory = layersImages.getTally().getMaxMemory();
+        long maxMemory = layersImages.getTally().getMaxMemory();
         if (maxMemory != Tally.UNLIMITED && Tally.computeImageSize(finalLegend) > maxMemory)
             throw new ServiceException(
                     LegendGraphicBuilder.MEMORY_USAGE_EXCEEDED,
